Pull patching methods up into linter
