+++ test/core/linter/linted_file_test.py
@@ -5,6 +5,7 @@ import logging
 import pytest
 
 from sqlfluff.core.linter import LintedFile
+from sqlfluff.core.linter.patch import FixPatch
 from sqlfluff.core.parser.markers import PositionMarker
 from sqlfluff.core.parser.segments import (
     BaseSegment,
@@ -12,7 +13,6 @@ from sqlfluff.core.parser.segments import (
     TemplateSegment,
 )
 from sqlfluff.core.parser.segments.raw import SourceFix
-from sqlfluff.core.rules.fix import FixPatch
 from sqlfluff.core.templaters import RawFileSlice, TemplatedFile
 from sqlfluff.core.templaters.base import TemplatedFileSlice
